+++ tests/fchk_parser_smoke.cpp
@@ -283,7 +283,7 @@ int main() {
         }
 
         const auto g = cov::parse_fchk(cart_g);
-        const std::vector<float> expected_g{
+        const std::vector<double> expected_g{
             15.0f, 5.0f, 1.0f, 14.0f, 13.0f, 9.0f, 4.0f, 6.0f, 2.0f,
             12.0f, 10.0f, 3.0f, 11.0f, 8.0f, 7.0f
         };
